+++ brain/spikeReportReader.cpp
@@ -34,6 +34,7 @@ public:
     {}
 
     brion::SpikeReport _report;
+    brion::Spikes _collected;
 };
 
 SpikeReportReader::SpikeReportReader( const brion::URI& uri )
@@ -46,13 +47,44 @@ SpikeReportReader::~SpikeReportReader()
     delete _impl;
 }
 
-brion::Spikes SpikeReportReader::getSpikes( const float startTime,
-                                            const float endTime )
+Spikes SpikeReportReader::getSpikes( const float startTime,
+                                     const float endTime )
 {
-    if(endTime <= startTime)
-        LBTHROW(std::logic_error("Start time should be strictly inferior to end time"));
-    _impl->_report.seek( startTime ).get();
-    return _impl->_report.readUntil( endTime ).get();
+    if( endTime <= startTime )
+        LBTHROW( std::logic_error(
+                     "Start time should be strictly inferior to end time" ));
+
+    if( _impl->_report.supportsBackwardSeek( ))
+    {
+        _impl->_report.seek( startTime ).get();
+        return _impl->_report.readUntil( endTime ).get();
+    }
+
+    // In reports that don't support seek we just want to move forward at
+    // least until end time, we'll find the window later.
+    // We use read instead of readUntil so the end time gets updated with
+    // the latest value possible. We also try to read always, even if all spikes
+    // in the requested window have been already collected.
+    auto& collected = _impl->_collected;
+    auto spikes = _impl->_report.read( endTime ).get();
+    if( !spikes.empty( ))
+    {
+        collected.reserve( collected.size() + spikes.size( ));
+        collected.insert( collected.end(), spikes.begin(), spikes.end( ));
+    }
+
+    return Spikes(
+        std::lower_bound( collected.begin(), collected.end(), startTime,
+                          []( const Spike& spike, const float val ){
+                              return spike.first < val; }),
+        std::lower_bound( collected.begin(), collected.end(), endTime,
+                          []( const Spike& spike, const float val ){
+                              return spike.first < val; }));
+}
+
+float SpikeReportReader::getEndTime() const
+{
+    return _impl->_report.getEndTime();
 }
 
 bool SpikeReportReader::hasEnded() const