+++ brion/plugin/spikeReportBluron.cpp
@@ -56,6 +56,9 @@ SpikeReportBluron::SpikeReportBluron( const SpikeReportInitData& initData )
     }
 
     _lastReadPosition = _spikes.begin();
+
+    if( !_spikes.empty( ))
+        _endTime = _spikes.rbegin()->first;
 }
 
 bool SpikeReportBluron::handles( const SpikeReportInitData& initData )